reto-03/gabrilov/requirements.txt
autopep8==1.7.0
mypy==0.971
mypy-extensions==0.4.3
python-magic==0.4.27
toml==0.10.2
types-toml==0.10.8
typing_extensions==4.3.0
